Add archiver cmake module
This allows you to attach an archiver task to a target. The
target will be copied into an archive/ directory on build
completion, tagged with the current commit id as returned
by git describe. The target archive directory can be
configured with the TARGET_ARCHIVE_DIR variable.

This is added to the <target>.fpga targets, to help reduce
the chances of accidentally removing the output from very
long fpga build processes.