Baseline: Make C++ API to no-tbx spotfinder
This makes it a lot easier to integrate, and also makes it a fair comparison
for benchmarking vs the C API - which added a lot of overhead (mainly from
a copy which can potentially be avoided).

Use this for Benchmarking, which also allows us to break out the
data-conversion copy which always happens inside dxtbx.