Add pipe output to spotfinder
Created a PipeHandler object in spotfinder.cc in order to manage 
thread-safe output of data. As well as including an argument for the 
pipe's file descriptor.

Added logic in service.py to find the spotfinder executable either
through an environment variable or in variants of a relative build path.

Created a thread in service.py in order to handle the concurrent output
of the spotfinder executable and subsequently send the data on to the
appropriate queue.

Created basic tests for the pipe-based output: pipetest.py and
piptest.sh.