Read trusted_range from data, and pass to kernel
Each image file has a trusted range value, and in DIALS this is applied at the
raw-data-reading stage. This expands h5read to get this data from HDF5, and
updates the other readers to pull equivalent information.

The CBF reader is incomplete, and so only the raw maximum value is used (not
read from the data files).